Rename 'trak' app to 'handler' app
    The handler app contains information about a waste handler, which is an instance of a RCRAinfo site on a manifest (e.g., signature info)
