+++ .idea/scopes/server.xml
@@ -1,3 +1,3 @@
 <component name="DependencyValidationManager">
-  <scope name="server" pattern="(file[haztrak]:server//*&amp;&amp;!file:__pycache__//*&amp;&amp;!file[haztrak]:server/.venv//*&amp;&amp;!file[haztrak]:server/.ruff_cache//*&amp;&amp;!file[haztrak]:server/.pytest_cache//*||file[haztrak]:server/.venv/*)&amp;&amp;!file[haztrak]:server/apps/trak/migrations//*||file[haztrak]:server/apps/trak//*" />
+  <scope name="server" pattern="(file[haztrak]:server//*&amp;&amp;!file:__pycache__//*&amp;&amp;!file[haztrak]:server/.venv//*&amp;&amp;!file[haztrak]:server/.ruff_cache//*&amp;&amp;!file[haztrak]:server/.pytest_cache//*||file[haztrak]:server/.venv/*)&amp;&amp;!file[haztrak]:server/apps/trak/migrations//*||file[haztrak]:server/apps/handler//*" />
 </component>
