+++ server/apps/core/management/commands/celery_beat.py
@@ -8,7 +8,7 @@ from django.core.management.base import BaseCommand
 from django.utils import autoreload
 from django_celery_beat.models import CrontabSchedule, IntervalSchedule, PeriodicTask
 
-from apps.trak.tasks import pull_federal_codes
+from apps.wasteline.tasks import pull_federal_codes
 
 CELERY_LOG_LEVEL = os.getenv("CELERY_LOG_LEVEL", "INFO")
 logger = logging.getLogger(__name__)
