+++ server/apps/handler/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TrakConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.trak"
+    name = "apps.handler"
