+++ server/apps/handler/models/signature.py
@@ -4,7 +4,7 @@ from typing import List, Literal, Optional
 
 from django.db import models
 
-from apps.trak.models.base_models import TrakBaseManager, TrakBaseModel
+from apps.handler.models.base_models import TrakBaseManager, TrakBaseModel
 
 logger = logging.getLogger(__name__)
 
@@ -106,7 +106,7 @@ class ESignature(TrakBaseModel):
     objects = ESignatureManager()
 
     manifest_handler = models.ForeignKey(
-        "trak.Handler",
+        "handler.Handler",
         related_name="e_signatures",
         on_delete=models.CASCADE,
     )
