+++ server/apps/handler/tests/services/test_manifest_services.py
