+++ server/apps/manifest/services/emanifest.py
@@ -8,11 +8,11 @@ from emanifest import RcrainfoResponse
 from requests import RequestException
 
 from apps.core.services import RcrainfoService, get_rcrainfo_client
+from apps.handler.models import QuickerSign
+from apps.handler.serializers import QuickerSignSerializer
 from apps.manifest.models import Manifest
 from apps.manifest.serializers import ManifestSerializer
 from apps.manifest.tasks import pull_manifest, sign_manifest
-from apps.trak.models import QuickerSign
-from apps.trak.serializers import QuickerSignSerializer
 
 logger = logging.getLogger(__name__)
