+++ server/apps/manifest/tests/test_models.py
@@ -4,9 +4,9 @@ import pytest
 from django.core.exceptions import ValidationError
 from django.db.models import Q
 
+from apps.handler.serializers import HandlerSerializer
 from apps.manifest.models import Manifest, draft_mtn, validate_mtn
 from apps.site.models import RcraSiteType
-from apps.trak.serializers import HandlerSerializer
 
 
 @pytest.mark.django_db
