+++ server/apps/manifest/views.py
@@ -7,6 +7,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from apps.handler.serializers import QuickerSignSerializer
 from apps.manifest.models import Manifest
 from apps.manifest.serializers import ManifestSerializer
 from apps.manifest.services import (
@@ -18,7 +19,6 @@ from apps.manifest.services import (
     update_manifest,
 )
 from apps.site.services import HaztrakSiteService
-from apps.trak.serializers import QuickerSignSerializer
 
 logger = logging.getLogger(__name__)
