--- server/apps/trak/views/__init__.py
@@ -1,7 +1,0 @@
-from apps.trak.views.manifest_view import (
-    ManifestViewSet,
-    MtnListView,
-    SaveElectronicManifestView,
-    SignManifestView,
-    SyncSiteManifestView,
-)
